+++ lib/uploadcare/resources/group_list.rb
@@ -7,7 +7,7 @@ module Uploadcare
       private
 
       def to_resource(api, group_data)
-        Uploadcare::Api::Group.new @api, group_data["id"], group_data
+        Uploadcare::Api::Group.new api, group_data["id"], group_data
       end
     end
   end